Modifications to orca system spec; added MPICXX flag
The MPICXX flag is necessary because the blitz configure script
does not properly deal with the sharcnet 'compile' scripting
command; since it does not actually contain 'icc' in the *NAME*
of the command, the configure seems to default to a
blitz/gnu/bzconfig.h scheme -- which is then not found during
bona fide compilation.  This is probably an underlying bug in
the blitz build system.