Re-ordered makefile libraries, putting fftw first
On systems using the MKL library, including the MKL (via $(LAPACK_LIB))
brings in the MKL fftw-wrapper bindings, rather than fftw itself.  This
is problematic on account of the fact that the MKL fftw wrapper doesn't
implement the parts of the library that SPINS actually uses; this leads
to an assertion failure on startup.