Included Options/-lboost_program_options at build
Also included an add_option speicalization for the string parameter
type.  It's convenient to allow for const char * as a default value for
these parameters, but that poses a problem for the templated add_option
function because of the lack of automatic conversion to const string.
By including a specialized overload of the template (with
implementation in Options.cpp), this problem is fixed.