Fixed bug relating to crash inside LAPACK for gmres
The rectangular-matrix solve inside the GMRES code is computed with the
LAPACK function DGELSD.  This function has a 'workspace query' in order
to give the proper allocation of the workspace array.  As it turns out,
the -integer- work array (as opposed to the floating-point one) is of a
fixed size, and it can be allocated ahead of time.  Not doing so caused
problems with MKL on SciNet, since even for the query the intel version
of LAPACK wrote to this array (causing a segmentation fault).