Modified the large bench_kh case to output memory usage
Using code lifted from internal_wave_dl, the benchmark case now reads
/proc/self/statm on a per-process basis in order to get the resident
set size.  This solution is probably -not portable- outside of Linux
systems; when the file is not present the execution will likely die with
an assertion failure.

Additionally, no effort is made to properly account for shared memory --
a hypothetical run that uses 1MB/proc of local memory and 99MB/proc of
shared memory will report a total of 100MB * procs memory usage, rather
than 1 * procs + 99MB.  A general solution here is difficult because it
relies on accounting for shared memory once-per-system (and nhost > 1
for cluster configurations).