Copied compiler-optimization flags to linker
In the system-specific build setting scripts, the OPTIM_LDFLAGS
variable was generally empty.  As it turns out, at least for
Itanium systems, this disabled a key optimization or two, leading
to the emergence of denormalized floats (which, on these systems,
are handled in software).  Testing shows that including -O3 at
the link step should keep this optimization enabled (really just
a modification of the default floating-point environment).  Since
this change is unlikely to break anything, the fix for all systems
has been to copy the $OPTIM_CFLAGS variable to $OPTIM_LDFLAGS.