Added build options for clean build on plata
Prior modifications to systems/plata.sh did not result in a clean
build; spins itself and the UMFPACK libraries were built in 64-
bit with the -m64 option, but fftw, blitz, and boost were not.
This led to build-time errors.  The fix involves making sure that
make_deps specifies CFLAGS/CXXFLAGS as appropriate, and this was
done via including $BOOST_OPTIONS/$FFTW_OPTIONS/etc flags.  Where
those didn't already exist in make_deps, they were added so that
it will be easy to make a similar change in the future for other
systems.