Add -ftz for builds on winisk
This patch specifically includes -ftz for winisk builds; this
forces denormalized floats to be flushed to 0, which should
avoid software-floating-point-faults that spam /var/log/messages
and probably slow down the program.  (Quick checks suggest a ~8%
speed increase for bench_dipole).

Additionally, make_deps is also extended to specifically include
CFLAGS and mark building with CC on the ./configure command line.
This should ensure that the library is built with the same
compiler (and settings) as SPINS itself.  If you have to modify
this part, pay special attention to the comment abour use of bash
array syntax.