Fixes to make old dipole cases compile
The old dipole cases in the cases/ directory (dipole crashing into
no-slip boundary) were initially written before the 'analysis'
routine was given pressure.  Adding pressure-analysis to NSIntegrator
changed the method's prototype, leading to a compile-time error on
these files.  The fix is to add a dummy 'pressure' reference to the
dipoles' analysis methods; the behaviour of the other parameters is
unchanged so no logic changes are necessary.