Cleaned up compiler warnings
As compiler versions march on, some previously-valid code was generating
warnings; these have been corrected.

*) In Plan_holder and T_util, using <ext/hash_map> was generating a
deprecation warning.  The proper replacement for the GCC <hash_map> is
the STL <unordered_map>, however that is part of the C++-0x standard
that is not yet supported by default in GCC 4.3-4.6, requiring a
compile-time command-line flag.

Since implementing *that* over a broad range of systems will be
extremely difficult, the better option for now is to silence the
deprecation warning by selectively undefining __DEPRECATED around the
inclusion of <ext/hash_map>.  This preserves any GCC deprecation
warnings around other bits of code.

*) In T_util and gmres_1d_solver, "literal string constants" were being
assigned to char *, when string literals are supposed to be const char
*.  This reflected an error in the variable spec; there's no reason that
the assigned-to-vars couldn't themselves be const char *.

As of this commit, the code compiles (make cases/bench_dipole.x)
compiles warning-free on belize.