wave_reader, tracer file output
Previously for the wave_reader case file, "tracer.0" and
"tracer_reader.m" files were being output regardless of whether the
configuration specified that there was a passive tracer.  This arose
because of a scoping error during tracer initalization; tracer.0
contained a copy of rho.0.

This has been fixed in this commit.  No numerical results were in error
because of this bug.