Create & impliment an easy to use diagnostic framework
Handling of the printing of diagnostics is now done in a few
short lines. This greatly simplifies the hard coded printing
previously done.

A diagnostic is saved by calling add_diagnostic:

add_diagnostic("Max_u",max_u,header,line);

Where "Max_u" is the name of the diagnostic (will be header in
the file). max_u is the value of the diagnostic. header and line
are the strings which hold the information of all headers and all
values for a given output, respectively.

Once all diagnostics are added the line is printed to the file
diagnostics.txt.